pyproject.toml
[project]
name = "velosim-mtl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mathieu Leduc-Hamel", email = "marrakis@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.2",
    "matplotlib>=3.10.8",
    "ollama>=0.6.1",
    "osmnx>=2.1.0",
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shapely>=2.1.2",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
]
